+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- fix PolygonIOCurrencySnapshotGramsFetcher and MetalPriceApiFetcher fetcher
 
 ## [7.29.6] - 2024-07-18
 ### Added